Move functionality from Network to API
