+++ src/lib/API.js
@@ -1,7 +1,120 @@
+import _ from 'underscore';
 import Ion from './Ion';
 import IONKEYS from '../IONKEYS';
-import {request} from './Network';
+import {request, setNewOfflineStatus, onReconnect} from './Network';
 import CONFIG from '../CONFIG';
+import * as Pusher from './Pusher/pusher';
+import ROUTES from '../ROUTES';
+
+let authToken;
+Ion.connect({key: IONKEYS.SESSION, path: 'authToken', callback: val => authToken = val});
+
+let currentUrl;
+Ion.connect({key: IONKEYS.CURRENT_URL, callback: url => currentUrl = url});
+
+/**
+ * When authTokens expire they will automatically be refreshed.
+ * The authorizer helps make sure that we are always passing the
+ * current valid token to generate the signed auth response
+ * needed to subscribe to Pusher channels.
+ */
+Pusher.registerCustomAuthorizer((channel, {authEndpoint}) => ({
+    authorize: (socketID, callback) => {
+        console.debug('[Network] Attempting to authorize Pusher');
+
+        const formData = new FormData();
+        formData.append('socket_id', socketID);
+        formData.append('channel_name', channel.name);
+        formData.append('authToken', authToken);
+
+        return fetch(authEndpoint, {
+            method: 'POST',
+            body: formData,
+        })
+            .then(authResponse => authResponse.json())
+            .then(data => callback(null, data))
+            .catch((err) => {
+                console.debug('[Network] Failed to authorize Pusher');
+                callback(new Error(`Error calling auth endpoint: ${err}`));
+            });
+    },
+}));
+
+// Initialize the pusher connection
+Pusher.init();
+
+/**
+ * Events that happen on the pusher socket are used to determine if the app is online or offline. The offline setting
+ * is stored in Ion so the rest of the app has access to it.
+ *
+ * @params {string} eventName,
+ * @params {object} data
+ */
+Pusher.registerSocketEventCallback((eventName, data) => {
+    let isCurrentlyOffline = false;
+    switch (eventName) {
+        case 'connected':
+            isCurrentlyOffline = false;
+            break;
+        case 'disconnected':
+            isCurrentlyOffline = true;
+            break;
+        case 'state_change':
+            if (data.current === 'connecting' || data.current === 'unavailable') {
+                isCurrentlyOffline = true;
+            }
+            break;
+        default:
+            break;
+    }
+    setNewOfflineStatus(isCurrentlyOffline);
+});
+
+/**
+ * Sets API data in the store when we make a successful "Authenticate"/"CreateLogin" request
+ *
+ * @param {object} data
+ * @param {string} exitTo
+ * @returns {Promise}
+ */
+function setSuccessfulSignInData(data, exitTo) {
+    let redirectTo;
+
+    if (exitTo && exitTo[0] === '/') {
+        redirectTo = exitTo;
+    } else if (exitTo) {
+        redirectTo = `/${exitTo}`;
+    } else {
+        redirectTo = ROUTES.HOME;
+    }
+
+    return Ion.multiSet({
+        // The response from Authenticate includes requestID, jsonCode, etc
+        // but we only care about setting these three values in Ion
+        [IONKEYS.SESSION]: _.pick(data, 'authToken', 'accountID', 'email'),
+        [IONKEYS.APP_REDIRECT_TO]: redirectTo,
+    });
+}
+
+// PUBLIC
+
+/**
+ * Register a callback function to be called when the network reconnects
+ *
+ * @public
+ * @param {function} cb
+ */
+function whenReconnected(cb) {
+    onReconnect(cb);
+}
+
+/**
+ * Get the authToken that the network uses
+ * @returns {string}
+ */
+function getAuthToken() {
+    return authToken;
+}
 
 /**
  * @param {object} parameters
@@ -23,6 +136,9 @@ function authenticate(parameters) {
         twoFactorAuthCode: parameters.twoFactorAuthCode,
         exitTo: parameters.exitTo,
     })
+        .then((response) => {
+            setSuccessfulSignInData(response, parameters.exitTo);
+        })
         .catch((err) => {
             console.error(err);
             console.debug('[SIGNIN] Request error');
@@ -32,13 +148,12 @@ function authenticate(parameters) {
 
 /**
  * @param {object} parameters
- * @param {string} parameters.authToken
  * @param {string} parameters.partnerUserID
  * @returns {Promise}
  */
 function deleteLogin(parameters) {
     return request('DeleteLogin', {
-        authToken: parameters.authToken,
+        authToken,
         partnerUserID: parameters.partnerUserID,
         partnerName: CONFIG.EXPENSIFY.PARTNER_NAME,
         partnerPassword: CONFIG.EXPENSIFY.PARTNER_PASSWORD,
@@ -49,4 +164,6 @@ function deleteLogin(parameters) {
 export {
     authenticate,
     deleteLogin,
+    getAuthToken,
+    whenReconnected,
 };
